pyproject.toml
[tool.poetry]
name = "serverless polars"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "3.10.10"
polars = ">=0.18.0"
duckdb = ">=0.8.0"
deltalake = ">=0.9.0"
ruff = "^0.0.270"
jupyterlab = "^4.0.0"
pandas = "^2.0.2"
ipykernel = "^6.23.1"
python-dotenv = "^1.0.0"
jupyter = "^1.0.0"
numpy = "^1.25.0"
s3fs = "^2023.6.0"
ipython = "^8.14.0"
patito = "^0.5.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
